+++ src/main/java/org/folio/inventory/support/InstanceUtil.java
@@ -39,8 +39,9 @@ public class InstanceUtil {
    * @return - result Instance
    */
   public static Instance mergeFieldsWhichAreNotControlled(Instance existing, org.folio.Instance mapped) {
-
     mapped.setId(existing.getId());
+    mapped.setDiscoverySuppress(existing.getDiscoverySuppress());
+    mapped.setStaffSuppress(existing.getStaffSuppress());
 
     List<ParentInstance> parentInstances = constructParentInstancesList(existing);