+++ NEWS.md
@@ -9,6 +9,7 @@
 * Adjust /mark-deleted endpoint behavior to set the "deleted" flag [MODINV-1138](https://folio-org.atlassian.net/browse/MODINV-1138)
 * Implement marc bib submatch [MODINV-1114](https://issues.folio.org/browse/MODINV-1114)
 * Actualize condition of deduplication logic for Authority creation [MODINV-1168](https://issues.folio.org/browse/MODINV-1168)
+* Update MARC-Instance mapping to account for Leader 05 value [MODINV-1140](https://issues.folio.org/browse/MODINV-1140)
 
 ## 21.0.0 2024-10-29
 * Existing "035" field is not retained the original position in imported record [MODINV-1049](https://folio-org.atlassian.net/browse/MODINV-1049)